M-cycles end where they're supposed to
Previously, extending a machine cycle had to be done with a single-cycle machine cycle called "extended". Also, internal M-cycles were limited to 3, 4, or 5 T-cycles.

This change changes the meaning of an extended cycle to indicate adding one T-cycle to the end of the current M-cycle. Also, the "internal" M-cycle is now just a single T-cycle that can be extended with "extended".

The result is that the "M-cycle done" signal now works as expected, which will make interrupts and bus requests easier to implement.